wip vote games in read

+++ game2Gather_Backend/src/main/java/de/szut/game2gather_backend/repository/VoteRepository.java
@@ -3,5 +3,5 @@ package de.szut.game2gather_backend.repository;
 import de.szut.game2gather_backend.entity.UserVote;
 import org.springframework.data.repository.CrudRepository;
 
-public interface VoteRepository extends CrudRepository<UserVote, Long> {
+public interface VoteRepository extends CrudRepository<UserVote, Integer> {
 }
